API_REST/info/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import InfoViewSet

router = DefaultRouter()
router.register(r'info', InfoViewSet, basename='info')

urlpatterns = [
    path('', include(router.urls)),
]